smaty-nfty2/requirements.txt
absl-py==0.8.1
adblockparser==0.7
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
aniso8601==8.0.0
app==0.0.1
appdirs==1.4.3
asgiref==3.3.4
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
astropy==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
Automat==20.2.0
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.7.1
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
blinker==1.4
bokeh==1.4.0
boto==2.49.0
boto3==1.12.24
botocore==1.15.24
Bottleneck==1.3.1
bs4==0.0.1
cachetools==4.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
charset-normalizer==2.0.4
Click==7.0
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.3
comtypes==1.1.7
conda==4.9.2
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
configparser==5.0.2
constantly==15.1.0
contextlib2==0.6.0.post1
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
cymem==1.31.2
Cython==0.29.15
cytoolz==0.9.0.1
dash==1.9.0
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
dask==2.10.1
dateparser==1.0.0
decorator==4.4.1
defusedxml==0.6.0
demjson==2.2.4
dill==0.2.9
distlib==0.3.0
distributed==2.10.0
docopt==0.6.2
docutils==0.15.2
dominate==2.4.0
emoji==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
Events==0.3
facebook-page-scraper==0.1.9
facebook-scraper==0.2.45
fake-useragent==0.1.11
falcon==2.0.0
fastapi==0.65.2
fastcache==1.1.0
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.1
Flask-API==2.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.8.0
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-User==1.0.2.2
Flask-WTF==0.14.3
fsspec==0.6.2
funcparserlib==0.3.6
funcy==1.14
future==0.17.1
gast==0.2.2
gensim==3.8.1
gevent==1.4.0
glob2==0.7
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-pasta==0.1.8
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.16.1
guess-language-spirit==0.5.3
h11==0.12.0
h2==3.2.0
h5py==2.10.0
HeapDict==1.0.1
hpack==3.0.0
html2text==2018.1.9
html5lib==1.0.1
hug==2.6.1
hyperframe==5.2.0
hyperlink==19.0.0
hypothesis==5.4.1
idna==2.8
imageio==2.6.1
imagesize==1.2.0
imbalanced-learn==0.6.2
imblearn==0.0
importlib-metadata==2.1.1
incremental==17.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jovian==0.2.26
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kaitaistruct==0.9
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
linkedin-scraper==2.8.2
llvmlite==0.31.0
locket==0.2.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
MechanicalSoup==1.1.0
mechanize==0.4.5
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.4.3
multipledispatch==0.6.0
murmurhash==0.28.0
mysql-connector==2.2.9
mysql-connector-python-rf==2.2.2
mysqlclient==1.4.6
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.6.8
numpy==1.15.4
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.1
pandas-datareader==0.8.1
pandocfilters==1.4.2
parse==1.19.0
parsel==1.5.2
parso==0.6.1
partd==1.1.0
passlib==1.7.2
path==13.1.0
pathlib2==2.3.5
patsy==0.5.1
pbr==5.4.4
pep8==1.7.1
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
plac==0.9.6
plotly==4.5.0
pluggy==0.13.1
ply==3.11
preshed==1.0.1
priority==1.3.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
Protego==0.1.16
protobuf==3.11.2
proxycrawl==3.1.0
psutil==5.6.7
py==1.8.1
pyaml==18.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.5
pydantic==1.8.2
PyDispatcher==2.0.5
pydivert==2.1.0
pyee==8.1.0
pyflakes==2.1.1
Pygments==2.5.2
PyHamcrest==2.0.2
PyJWT==2.1.0
pyLDAvis==2.1.2
pylint==2.4.4
PyMySQL==0.9.3
pyodbc===4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyppeteer==0.2.5
pyquery==1.4.3
pyreadline==2.1
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qt5reactor==0.6
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
queuelib==1.5.0
readability-lxml==0.7
regex==2018.7.11
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.16.0
rsa==4.0
ruamel_yaml==0.15.87
s3transfer==0.3.3
schedule==0.6.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.1.0
Scrapy==2.0.1
scrapy-splash==0.7.2
scrapyd==1.2.1
scrapyd-client==1.2.0a1
scrapyrt==0.11.0
seaborn==0.10.0
selenium==3.141.0
selenium-wire==4.3.1
Send2Trash==1.5.0
sentencepiece==0.1.94
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
smart-open==1.10.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
spacy==2.0.12
speaklater==1.3
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
splash==3.4.1
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.13
starlette==0.14.2
stats==0.1.2a0
statsmodels==0.11.0
stevedore==1.32.0
sympy==1.5.1
tables==3.6.1
TBB==0.1
tblib==1.6.0
tensorboard==1.15.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
thinc==6.10.3
toolz==0.10.0
torch @ https://download.pytorch.org/whl/cpu/torch-0.4.1-cp37-cp37m-win_amd64.whl
torchtext==0.6.0
torchvision @ https://download.pytorch.org/whl/torchvision-0.1.6-py3-none-any.whl
tornado==6.0.3
tqdm==4.42.1
trackthenews==0.1.10
traitlets==4.3.3
Twisted==20.3.0
twitter-scraper==0.4.4
twython==3.7.0
typing-extensions==3.10.0.0
tzlocal==2.0.0
ujson==1.35
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.26.6
uuid==1.30
uvicorn==0.14.0
virtualenv==20.0.4
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
visitor==0.1.3
w3lib==1.21.0
wcwidth==0.1.8
webdriver-manager==3.2.2
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
wsproto==1.0.0
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.7
xlwings==0.17.1
xlwt==1.3.0
xvfbwrapper==0.2.9
youtube-search==2.1.0
zict==1.0.0
zipp==2.2.0
zope.interface==5.0.2
